dialogs added

+++ Projects/App/wwwroot/system/landing.controller.js
@@ -5,13 +5,22 @@
          .module('app')
          .controller('LandingController', LandingController);
 
-    LandingController.$inject = ['userReportsService'];
+    LandingController.$inject = ['userReportsService', 'ngDialog'];
 
-    function LandingController(userReportsService) {
+    function LandingController(userReportsService, ngDialog) {
         var vm = this;
         vm.reports = userReportsService.reports;
         vm.remove = function (report) {
-            userReportsService.remove(report.Id);
+            var warningDialog = ngDialog.open({
+                template: 'wwwroot/report/popups/remove.warning.template.html'
+            });
+
+            warningDialog.closePromise.then(function (data) {
+                if (data.value === true) {
+                    report.isRemoving = true;
+                    userReportsService.remove(report.Id);
+                }
+            });            
         }
     }
 })();
